Break down domainData reducer
