--- frontend/src/common/action-types/domainData.js
@@ -1,34 +1,0 @@
-export const SET_USER_INFORMATION = 'domain-data/SET_USER_INFORMATION';
-export const UNSET_USER = 'domain-data/UNSET_USER';
-export const SET_USERS_INFORMATION = 'domain-data/SET_USERS_INFORMATION';
-export const SET_USER_PROJECTS = 'domain-data/SET_USER_PROJECTS';
-export const SET_USER_PROJECT = 'domain-data/SET_USER_PROJECT';
-export const SET_USER_PROJECT_OPTIONS = 'domain-data/SET_USER_PROJECT_OPTIONS';
-export const SET_USERS_PROJECT_MEMBERSHIP = 'domain-data/SET_USERS_PROJECT_MEMBERSHIP';
-export const SET_USER_PROJECT_MEMBERSHIP = 'domain-data/SET_USER_PROJECT_MEMBERSHIP';
-export const UNSET_USER_PROJECT_MEMBERSHIP = 'domain-data/UNSET_USER_PROJECT_MEMBERSHIP';
-export const UNSET_USER_PROJECT = 'domain-data/UNSET_USER_PROJECT';
-export const SET_USER_GROUPS = 'domain-data/SET_USER_GROUPS';
-export const SET_USER_GROUP = 'domain-data/SET_USER_GROUP';
-export const UNSET_USER_GROUP = 'domain-data/UNSET_USER_GROUP';
-export const SET_USERS_MEMBERSHIP = 'domain-data/SET_USERS_MEMBERSHIP';
-export const SET_USER_MEMBERSHIP = 'domain-data/SET_USER_MEMBERSHIP';
-export const UNSET_USER_MEMBERSHIP = 'domain-data/UNSET_USER_MEMBERSHIP';
-
-export const UNSET_REGION = 'domain-data/UNSET_REGION';
-export const ADD_NEW_REGION = 'domain-data/ADD_NEW_REGION';
-export const REMOVE_PROJECT_REGION = 'domain-data/REMOVE_PROJECT_REGION';
-
-export const SET_LEAD_FILTER_OPTIONS = 'domain-data/SET_LEAD_FILTER_OPTIONS';
-
-export const SET_REGIONS = 'domain-data/SET_REGIONS';
-export const SET_REGION_DETAILS = 'domain-data/SET_REGION_DETAILS';
-export const SET_ADMIN_LEVELS_FOR_REGION = 'domain-data/SET_ADMIN_LEVELS_FOR_REGION';
-export const ADD_ADMIN_LEVEL_FOR_REGION = 'domain-data/ADD_ADMIN_LEVEL_FOR_REGION';
-export const UNSET_ADMIN_LEVEL_FOR_REGION = 'domain-data/UNSET_ADMIN_LEVEL_FOR_REGION';
-
-export const SET_ANALYSIS_FRAMEWORK = 'domain-data/SET_ANALYSIS_FRAMEWORK';
-export const SET_ANALYSIS_FRAMEWORKS = 'domain-data/SET_ANALYSIS_FRAMEWORKS';
-export const ADD_NEW_AF = 'domain-data/ADD_NEW_AF';
-export const SET_PROJECT_AF = 'domain-data/SET_PROJECT_AF';
-export const SET_AF_DETAIL = 'domain-data/SET_AF_DETAIL';
