+++ frontend/src/common/reducers/siloDomainData/common.js
@@ -1,7 +1,7 @@
 import update from '../../../public/utils/immutable-update';
 
 import { LOGOUT_ACTION } from '../../reducers/auth';
-import { SET_USER_PROJECTS } from '../../action-types/domainData';
+import { SET_USER_PROJECTS } from '../../reducers/domainData/projects';
 
 import initialSiloDomainData from '../../initial-state/siloDomainData';
