+++ frontend/src/common/reducers/siloDomainData/common.test.js
@@ -1,5 +1,5 @@
 import { LOGOUT_ACTION } from '../../reducers/auth';
-import { SET_USER_PROJECTS } from '../../action-types/domainData';
+import { SET_USER_PROJECTS } from '../../reducers/domainData/projects';
 
 import initialSiloDomainData from '../../initial-state/siloDomainData';
 import reducers, {
