+++ frontend/src/common/redux.js
@@ -1,10 +1,10 @@
-export * from './action-creators/domainData';
-
 export * from './selectors/auth';
 export * from './selectors/domainData';
 export * from './selectors/siloDomainData';
 
+// NOTE: named exports from reducers are action-creators and action-types
 export * from './reducers/auth';
+
 export * from './reducers/siloDomainData/common';
 export * from './reducers/siloDomainData/leads';
 export * from './reducers/siloDomainData/entries';
@@ -12,3 +12,10 @@ export * from './reducers/siloDomainData/addLeads';
 export * from './reducers/siloDomainData/editEntries';
 export * from './reducers/siloDomainData/analysisFramework';
 export * from './reducers/siloDomainData/categoryEditor';
+
+export * from './reducers/domainData/analysisFrameworks';
+export * from './reducers/domainData/leadFilter';
+export * from './reducers/domainData/projects';
+export * from './reducers/domainData/regions';
+export * from './reducers/domainData/userGroups';
+export * from './reducers/domainData/users';
